+++ virtuo-piano/src/app/page.tsx
@@ -1,104 +1,17 @@
 import { redirect } from 'next/navigation';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/lib/authoption';
-import Image from 'next/image';
 import prisma from '@/lib/prisma';
-import { User } from '@/common/types';
-import SignOutButton from '@/components/SignOutButton';
 
-async function getUsers(): Promise<User[]> {
-  const users = await prisma.user.findMany({
-    select: {
-      id: true,
-      userName: true,
-      level: true,
-      createdAt: true,
-    },
-  });
-  return users;
-}
+import LandingPage from './landingPage/page';
+import { User } from '@/common/types';
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
 
   if (!session) {
-    redirect('/auth/signup');
+    redirect('/auth/register');
   }
 
-  const users = await getUsers();
-
-  return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <div className="w-full flex justify-end">
-        <SignOutButton />
-      </div>
-      <main className="flex flex-col gap-[32px] row-start-2 items-center w-full max-w-4xl">
-        <h1 className="text-2xl font-bold mb-4">Liste des Utilisateurs</h1>
-
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 w-full">
-          {users.map((user: User) => (
-            <div
-              key={user.id}
-              className="bg-white dark:bg-gray-800 p-4 rounded-lg shadow-md"
-            >
-              <h2 className="text-xl font-semibold mb-2">{user.userName}</h2>
-              <p className="text-gray-600 dark:text-gray-300">
-                Niveau: {user.level}
-              </p>
-              <p className="text-sm text-gray-500 dark:text-gray-400">
-                Inscrit le: {new Date(user.createdAt).toLocaleDateString()}
-              </p>
-            </div>
-          ))}
-        </div>
-      </main>
-      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
-    </div>
-  );
+  return <LandingPage />;
 }
