--- virtuo-piano/src/features/auth/LoginForm.tsx
@@ -1,54 +1,0 @@
-'use client';
-
-import { useState } from 'react';
-import { signIn } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-
-export default function LoginForm() {
-  const router = useRouter();
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    const res = await signIn('credentials', {
-      email,
-      password,
-      redirect: false,
-    });
-
-    if (res?.error) {
-      setError('Identifiants invalides');
-    } else {
-      router.push('/'); // ou dashboard
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="space-y-4">
-      <input
-        type="email"
-        placeholder="Email"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-        className="w-full p-2 border rounded"
-      />
-      <input
-        type="password"
-        placeholder="Mot de passe"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-        className="w-full p-2 border rounded"
-      />
-      <button
-        type="submit"
-        className="w-full bg-blue-600 text-white p-2 rounded"
-      >
-        Se connecter
-      </button>
-      {error && <p className="text-red-500 text-sm">{error}</p>}
-    </form>
-  );
-}
