+++ virtuo-piano/src/lib/authoption.ts
@@ -19,7 +19,7 @@ export const authOptions: NextAuthOptions = {
     strategy: 'jwt',
   },
   pages: {
-    signIn: '/auth/signin',
+    signIn: '/auth/login',
   },
   providers: [
     CredentialsProvider({
